pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "anvilcv"
version = "0.1.0"
description = "Developer-native, AI-powered resume engine — a fork of rendercv"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Anvil Contributors" },
]
keywords = ["resume", "cv", "ats", "ai", "typst", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Text Processing :: Markup",
]
dependencies = [
    # Inherited from rendercv
    "pydantic[email]>=2.0",
    "pydantic-extra-types>=2.0",
    "typer>=0.12",
    "rich>=13.0",
    "Jinja2>=3.1",
    "ruamel.yaml>=0.18",
    "watchdog>=4.0",
    "packaging>=23.0",
    "phonenumbers>=8.0",
    "markdown>=3.0",
    "annotated-doc>=0.0.4",
    "rendercv-fonts>=0.5",
    # Anvil additions
    "pdfminer.six>=20231228",
    "readability-lxml>=0.8",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-xdist>=3.5",
    "mypy>=1.8",
    "ruff>=0.3",
    "vcrpy>=6.0",
    "responses>=0.25",
]
ai = [
    "anthropic>=0.25",
    "openai>=1.12",
]
full = [
    "anvilcv[ai]",
    "typst>=0.11",
]

[project.scripts]
anvil = "anvilcv.cli.entry_point:main"

[tool.ruff]
target-version = "py312"
line-length = 100
extend-exclude = ["src/anvilcv/vendor/"]

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "nightly: marks tests that run nightly against live APIs",
]
